Updated README.md
Welcome to The Archive! This is where I plan on storing all of my old code that isn't tied to any specific project I am currently working on. Old personal projects, iOS school code, web themes and so much more!

I will be updating this repository as time goes on so keep checking back for more of my old code.

Thank you for reading and have a blessed day!
.j.